pyproject.toml
[tool.poetry]
name = "s3-presign-upload"
version = "0.1.0"
description = ""
authors = ["Ratchaphon1412 <Ratchaphon.h111@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^5.0.7"
djangorestframework = "^3.15.2"
django-environ = "^0.11.2"
google-cloud-storage = "^2.18.0"
six = "^1.16.0"
google-auth = "^2.33.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
